+++ packages/react/src/components/Menu/MenuDivider.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react'
 import * as PropTypes from 'prop-types'
+import { Accessibility } from '../../lib/accessibility/types'
+import { menuDividerBehavior } from '../../lib/accessibility'
 
 import {
   createShorthandFactory,
@@ -11,6 +13,12 @@ import {
 import { ReactProps } from '../../../../../types/utils'
 
 export interface MenuDividerProps extends UIComponentProps, ColorComponentProps {
+  /**
+   * Accessibility behavior if overridden by the user.
+   * @default menuDividerBehavior
+   */
+  accessibility?: Accessibility
+
   vertical?: boolean
   primary?: boolean
   secondary?: boolean
@@ -28,17 +36,25 @@ class MenuDivider extends UIComponent<ReactProps<MenuDividerProps>, any> {
 
   static defaultProps = {
     as: 'li',
+    accessibility: menuDividerBehavior as Accessibility,
   }
 
   static propTypes = {
     ...commonPropTypes.createCommon({ content: false, children: false, color: true }),
+    accessibility: PropTypes.func,
     primary: PropTypes.bool,
     secondary: PropTypes.bool,
     vertical: PropTypes.bool,
   }
 
-  renderComponent({ ElementType, classes, unhandledProps }) {
-    return <ElementType {...unhandledProps} className={classes.root} />
+  renderComponent({ ElementType, classes, unhandledProps, accessibility }) {
+    return (
+      <ElementType
+        {...accessibility.attributes.root}
+        {...unhandledProps}
+        className={classes.root}
+      />
+    )
   }
 }